Move package `metricssql` under package `metricsview`
